pyproject.toml
[project]
name = "try-smolagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=8.22.1",
    "python-dotenv>=1.1.0",
    "smolagents[telemetry]>=1.13.0",
]
